% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-credentials.R
\name{az_get_env_credentials}
\alias{az_get_env_credentials}
\title{Fetch Azure credentials stored as environmental variables, if they exist}
\usage{
az_get_env_credentials()
}
\value{
A list containing all required credentials or an informative error
if any are missing
}
\description{
Fetch Azure credentials stored as environmental variables, if they exist
}
\details{
See \href{https://github.com/CDCgov/cfazuR/tree/main?tab=readme-ov-file#required-setup-after-installation}{Required credential setup} #nolint
for help finding and specifying Azure credntials.
}
