% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-credentials.R
\name{az_required_credentials}
\alias{az_required_credentials}
\title{List names of all required azure credentials}
\usage{
az_required_credentials()
}
\value{
a character vector containing all required credentials
}
\description{
List names of all required azure credentials
}
\details{
used to view required credentials and internally to ensure the set
of required credentials is consistent in downstream functions
}
