% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-credentials.R
\name{az_set_env_vars}
\alias{az_set_env_vars}
\title{Helper fuction to set Azure credentials as environment variables}
\usage{
az_set_env_vars(
  AZURE_CLIENT_ID = "",
  AZURE_TENANT_ID = "",
  AZURE_CLIENT_SECRET = "",
  AZURE_SUBSCRIPTION_ID = "",
  AZURE_RESOURCE_GROUP = "",
  AZURE_STORAGE_ACCOUNT = ""
)
}
\arguments{
\item{AZURE_CLIENT_ID}{A character vector, the Azure client ID}

\item{AZURE_TENANT_ID}{A character vector, the Azure tenant ID}

\item{AZURE_CLIENT_SECRET}{A character vector, the Azure client secret}

\item{AZURE_SUBSCRIPTION_ID}{A character vector, the Azure subscription ID
@param AZURE_RESOURCE_GROUP A character vector, the Azure resource group}

\item{AZURE_STORAGE_ACCOUNT}{A character vector, the Azure storage account}
}
\description{
Helper fuction to set Azure credentials as environment variables
}
\details{
This function sets transient environment variables for the current R session.
To set these variables permanently, add them to your .Renviron file using
the syntax: \code{Sys.setenv(AZURE_SUBSCRIPTION_ID = "your_subscription_id")}.
Try usethis::edit_r_environ() to open your .Renviron file in RStudio.
For security reasons, never add the client ID or client secret to your
.Renviron file.
}
