% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-credentials.R
\name{az_validate_credlist}
\alias{az_validate_credlist}
\title{Checks that a list of credentials contains all expected credentials}
\usage{
az_validate_credlist(cred_list)
}
\arguments{
\item{cred_list}{a named list of credentials. Names must match the set of
required credentials returned by \code{\link[=az_required_credentials]{az_required_credentials()}}. All values must
be non-empty strings.}
}
\description{
Checks that a list of credentials contains all expected credentials
}
\details{
Checks that credentials exist, but does not connect to Azure to
check that the credentials are valid
}
