% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-credentials.R
\name{fetch_env_credential}
\alias{fetch_env_credential}
\title{Internal function: fetch one Azure credential from environment variable
and throw an informative error if credential is not found}
\usage{
fetch_env_credential(env_var)
}
\arguments{
\item{env_var}{A character vector, the credential(s) to fetch}
}
\value{
The value stored as the environment variable "env_var" if it exists
}
\description{
Internal function: fetch one Azure credential from environment variable
and throw an informative error if credential is not found
}
